fix(ci): bound, isolate and sanitise the gate probe's shell
Four findings, each reproduced first. Three produced a visibly wrong
verdict — an unallowlisted alias classified `root-tooling-scripts`.

A nonterminating classifier hung the routing test until whatever outer
limit the job had. Every probe shell now has a 30s deadline: the classifier
takes 130ms and the extraction 75ms against the real gate, so 30s is ~200x
the slower of the two, and a `while :; do :; done` cannot finish inside it.
Timing out is reported as "did not terminate", not as a job timeout.

`spawnSync`'s own timeout only signals the shell it started, and the
classifier runs inside a `$(…)`: killing the parent left two processes
spinning at 43% CPU, reparented to init. Measured, not assumed. The shell
now starts `detached` and the timeout kills the whole process group.

Non-interactive bash sources $BASH_ENV and imports exported functions at
STARTUP, before the probe empties PATH or installs the guard, so whatever
they define is a function `command -v` then reports as available. The
child now gets a sanitised environment: no BASH_ENV, ENV, BASH_FUNC_*,
SHELLOPTS or BASHOPTS.

`"new_helper"` is an ordinary call the left-boundary match missed, and in
a branch no synthetic path enters the runtime guard never fired either.
The helper search now also runs over a copy with quotes and backslashes
removed.

Ambient filesystem reads are closed structurally rather than by naming the
primitives that can make them — `[[ -f ]]`, `test -e`, globs, `compgen -G`
is a list without an end. The classifier runs from a fresh empty
directory, so the checkout is not reachable at all. The real gate
classifier still verdicts correctly from there, which is the case that
would have broken; a test pins it on every installed bash.

The probe's own invariants move to
check-sentry-suites-in-ci-gate-probe.test.mjs, since the lifecycle file
crossed the 1,000-line cap.

Refs #1779
